added javafx app container for standalone webapp
The webview of javafx is now used to run a native browser window
that connects to the webapp on localhost, effectively making it
a local application. The jetty server is saved into an atom by
the application and stopped when the window is closed.

The app.clj file is well reusable into other projects to turn a
webapp into a local desktop application.